useClickOutside refactored to use callbacks. Note and headers components broken down smaller
