+++ frontend/src/components/main/NoteComponents/NoteStyles.module.css
@@ -28,7 +28,7 @@
   resize: none;
   overflow: hidden;
   max-height: min-content;
-  padding: .5rem .5rem 0 .5rem;
+  padding: .5rem .5rem .5rem .5rem;
   color: inherit
 }
