+++ frontend/src/hooks/useClickOutside.tsx
@@ -1,27 +1,22 @@
-import { Dispatch, RefObject, SetStateAction, useEffect } from 'react'
+import { RefObject, useEffect } from 'react'
 
 
 
 const useClickOutside = (
   ref: RefObject<HTMLElement>,
-  states: boolean[],
-  setters: Dispatch<SetStateAction<boolean>>[]
+  handler: () => void
 ): void => {
 
   useEffect(() => {
     const handleClickOutside = (event) => {
       if (ref.current && !ref.current.contains(event.target as Node)) {
         event.stopPropagation()
-        if (states.some(state => state)) {
-          setters.forEach(setState => {
-            setState(false)
-          })
-        }
+        handler()
       }
     }
     document.addEventListener("click", handleClickOutside, true)
 
     return () => document.removeEventListener("click", handleClickOutside, true)
-  }, [states, ref, setters])
+  }, [ref, handler])
 }
 export default useClickOutside
